Refactor services and utilities into core app
- Moved Azure Blob storage functions to core/services/storage.py
- Moved file upload service to core/services/uploads.py
- Moved import services for study identifiers and clinical data to core/services/imports.py
- Moved clinical data retrieval functions to core/clinical.py
- Moved CSV export functions to core/utils/export.py
- Moved queryset filtering functions to core/utils/queries.py
- Moved dataframe manipulation functions to core/utils/dataframes.py
- Moved historical changes tracking functions to core/utils/history.py
- Updated imports across the application to reflect new module structure
- Enhanced maintainability and modularity of the codebase
