+++ codex_recommendations.md
@@ -12,29 +12,7 @@
 - Introduce class-based views or reusable mixins (e.g. `PaginatedFilterView`) that handle the boilerplate for pagination/querystring reconstruction and CSV export endpoints. This will shrink view modules and centralise behaviour changes.
 - Keep API views separate from template-rendering views. Today REST endpoints and HTML handlers live side-by-side in the same module (`datasets/views.py`), which complicates authentication/permission policies.
 
-## 3. Refactor service and utility layers into cohesive modules
-
-- `app/services.py` contains unrelated responsibilities: Azure Blob storage helpers, file upload coordination, and Study Identifier import logic (`app/services.py:14-200`).
-- `app/utils.py` likewise blends CSV export, queryset filtering, and pandas dataframe shaping (`app/utils.py:11-200`).
-- Break these into purpose-driven modules, e.g. `services/storage.py`, `services/imports.py`, and `utils/export.py`/`utils/dataframes.py`. Define clear public interfaces so other apps consume narrowly-scoped helpers, making it easier to swap storage backends or extend import pipelines.
-
-## 4. Align templates with their owning app and introduce shared layout components
-
-- Many templates live at the project root (`templates/management.html`, `templates/reference.html`) despite belonging to specific domains handled in the `app` package, increasing coupling between apps and template names.
-- Mirror the Django app structure inside `templates/` so that URLs in, for example, the future `samples` app resolve to `samples/…` templates. Reserve the project-level `templates/` root for global layouts (`layouts/`), partials (`includes/`), and error pages.
-- Extract repeated dashboard scaffolding (navigation, cards) into reusable includes or a base template so new apps can inherit consistent layouts without copy/paste.
-
-## 5. Clarify settings layering and environment configuration
-
-- Base settings import `.env` unconditionally (`config/settings/base.py`), yet production/local modules override numerous environment-specific values. Move environment-only defaults (e.g. debug emails, storages, Azure credentials) into `local.py`/`production.py` and guard `.env` loading so production can rely on environment variables.
-- Consider adopting `config/settings/__init__.py` that selects a settings module via `DJANGO_SETTINGS_MODULE` and document the expected environment variables. This keeps the base layer environment-agnostic and eases deployment automation.
-
-## 6. Separate dataset orchestration concerns
+## 3. Separate dataset orchestration concerns
 
 - Dataset HTML views, API serializers, and access tracking logic live in a single module (`datasets/views.py`), making it harder to evolve API contracts independently of the dashboard.
 - Split into `datasets/api/views.py` and `datasets/web/views.py`, and add a thin service layer to encapsulate history tracking and export logic (`datasets/utils.py`) so both entry points reuse the same orchestration code.
-
-## 7. Strengthen test organisation after restructuring
-
-- Tests currently mirror the monolithic layout (`app/tests/...`). As you extract domain apps, create per-app `tests/` packages and move fixtures/factories alongside domain code (`datasets/factories.py`, `app/factories.py`).
-- Promote shared fixtures (e.g. user factories, permission helpers) into a project-level `tests/common/` module to avoid cross-app imports.
