+++ core/services/__init__.py
@@ -1,0 +1,3 @@
+# /Users/chershiongchuah/Developer/musicsamples/core/services/__init__.py
+# This file initializes the core.services package.
+# It provides a namespace for service-related modules in the core app.
