+++ core/services/imports.py
@@ -1,144 +1,11 @@
-import logging
-from datetime import datetime, timedelta
-from typing import Optional
+# /Users/chershiongchuah/Developer/musicsamples/core/services/imports.py
+# This module provides import services for study identifiers and clinical data.
+# It processes data imports from dataframes and ensures data integrity.
 
 import pandas as pd
-from azure.storage.blob import BlobSasPermissions, BlobServiceClient, generate_blob_sas
-from django.conf import settings
-from django.db import models, transaction
-from django.utils import timezone
+from django.db import transaction
 
-from app.models import ClinicalData, DataStore, StudyIdentifier, file_generate_name, file_upload_path
-
-
-def azure_get_blob_service_client():
-    return BlobServiceClient(
-        account_url=f"https://{settings.AZURE_ACCOUNT_NAME}.blob.core.windows.net",
-        credential=settings.AZURE_ACCOUNT_KEY,
-    )
-
-
-def azure_generate_download_link(file, download=False):
-    """
-    Pass in the file object stored on Azure Blob Storage.
-    The blob path is constructed as <study_name>/<category>/<filename>.
-    Generates a SAS token valid for 1 hour with read permission and,
-    if download is True, an attachment disposition.
-    """
-    container_name = settings.AZURE_CONTAINER_NAME
-    blob_name = f"{file.file.name}"
-
-    expiration_time = datetime.utcnow() + timedelta(hours=1)
-
-    try:
-        sas_kwargs = {
-            "account_name": settings.AZURE_ACCOUNT_NAME,
-            "container_name": container_name,
-            "blob_name": blob_name,
-            "account_key": settings.AZURE_ACCOUNT_KEY,
-            "permission": BlobSasPermissions(read=True),
-            "expiry": expiration_time,
-        }
-        if download:
-            sas_kwargs["content_disposition"] = "attachment"
-
-        sas_token = generate_blob_sas(**sas_kwargs)
-    except Exception as e:
-        logging.error(e)
-        return None
-
-    download_url = (
-        f"https://{settings.AZURE_ACCOUNT_NAME}.blob.core.windows.net/{container_name}/{blob_name}?{sas_token}"
-    )
-    return download_url
-
-
-def azure_delete_file(file):
-    """
-    Pass in the file
-    """
-    blob_service_client = azure_get_blob_service_client()
-    container_name = settings.AZURE_CONTAINER_NAME
-    blob_name = f"{file.file.name}"
-
-    try:
-        blob_client = blob_service_client.get_blob_client(container=container_name, blob=blob_name)
-        blob_client.delete_blob()
-    except Exception as e:
-        logging.error(e)
-        return None
-    return True
-
-
-class FileDirectUploadService:
-    @transaction.atomic
-    def start(
-        self,
-        category: str,
-        study_name: str,
-        file_name: str,
-        music_timepoint: str,
-        marvel_timepoint: str,
-        comments: str,
-        sampling_date: Optional[datetime] = None,
-        study_id: Optional[str] = None,
-        uploaded_by=None,
-    ) -> dict:
-        if study_id:
-            study_id = study_id.upper()
-            study_identifier, _ = StudyIdentifier.objects.get_or_create(name=study_id)
-            formatted_file_name = file_generate_name(file_name, study_name, study_id)
-        else:
-            formatted_file_name = file_generate_name(file_name, study_name)
-            study_identifier = None
-
-        file = DataStore(
-            category=category,
-            study_name=study_name,
-            music_timepoint=music_timepoint,
-            marvel_timepoint=marvel_timepoint,
-            sampling_date=sampling_date,
-            comments=comments,
-            study_id=study_identifier,
-            file_type=file_name.split(".")[-1],
-            original_file_name=file_name,
-            formatted_file_name=formatted_file_name,
-            uploaded_by=uploaded_by,
-            file=None,
-        )
-        file.full_clean()
-        file.save()
-
-        upload_path = file_upload_path(file, formatted_file_name)
-
-        file.file = file.file.field.attr_class(file, file.file.field, upload_path)
-        file.save()
-
-        try:
-            sas_kwargs = {
-                "account_name": settings.AZURE_ACCOUNT_NAME,
-                "container_name": settings.AZURE_CONTAINER_NAME,
-                "blob_name": upload_path,
-                "account_key": settings.AZURE_ACCOUNT_KEY,
-                "permission": BlobSasPermissions(write=True),
-                "expiry": timezone.now() + timedelta(hours=1),
-            }
-
-            sas_token = generate_blob_sas(**sas_kwargs)
-            upload_url = f"https://{settings.AZURE_ACCOUNT_NAME}.blob.core.windows.net/{settings.AZURE_CONTAINER_NAME}/{upload_path}?{sas_token}"
-
-        except Exception as e:
-            logging.error(e)
-            return {"error": "Failed to generate upload URL"}
-        return {"id": file.pk, "upload_url": upload_url}
-
-    @transaction.atomic
-    def finish(self, *, file: DataStore) -> DataStore:
-        file.upload_finished_at = timezone.now()
-        file.full_clean()
-        file.save()
-
-        return file
+from app.models import ClinicalData, StudyIdentifier
 
 
 class StudyIdentifierImportService:
@@ -371,113 +238,3 @@ class ClinicalDataImportService:
                 skipped += 1
 
         return {"created": created, "updated": updated, "skipped": skipped, "errors": errors if errors else None}
-
-
-def get_sample_with_clinical_data(sample):
-    """
-    Get clinical data for a specific sample based on study type-specific matching
-    """
-    from app.models import ClinicalData
-
-    # First check if this sample has a study_id
-    if not sample.study_id:
-        return None
-
-    study_name = sample.study_id.study_name.lower() if sample.study_id.study_name else ""
-    sample_date = sample.sample_datetime.date() if sample.sample_datetime else None
-
-    try:
-        # Different lookup strategy based on study name
-        if "gidamps" in study_name:
-            # GI-DAMPs uses study_id and sample_date
-            if not sample_date:
-                return None
-
-            clinical_data = ClinicalData.objects.get(study_id=sample.study_id, sample_date=sample_date)
-
-        elif "music" in study_name:
-            # MUSIC and Mini-MUSIC use study_id and music_timepoint
-            if not sample.music_timepoint:
-                # Try sample_date as fallback if available
-                if sample_date:
-                    clinical_data = ClinicalData.objects.get(study_id=sample.study_id, sample_date=sample_date)
-                else:
-                    return None
-            else:
-                clinical_data = ClinicalData.objects.get(
-                    study_id=sample.study_id, music_timepoint=sample.music_timepoint
-                )
-
-        else:
-            # Default to sample_date for unknown studies
-            if not sample_date:
-                return None
-
-            clinical_data = ClinicalData.objects.get(study_id=sample.study_id, sample_date=sample_date)
-
-        return clinical_data
-
-    except ClinicalData.DoesNotExist:
-        return None
-
-
-def get_samples_with_clinical_data(queryset):
-    """
-    Annotate a queryset of samples with clinical data using study-specific matching
-    """
-    from django.db.models import Case, OuterRef, Subquery, When
-    from django.db.models.functions import Lower
-
-    from app.models import ClinicalData
-
-    # First, annotate with sample_date for joining
-    samples_with_date = queryset.annotate(
-        sample_date=models.functions.TruncDate("sample_datetime"), study_name_lower=Lower("study_id__study_name")
-    )
-
-    # Create a conditional subquery based on study name
-    # For GI-DAMPs - use study_id and sample_date
-    gidamps_clinical_data = ClinicalData.objects.filter(
-        study_id=OuterRef("study_id"), sample_date=OuterRef("sample_date")
-    )
-
-    # For MUSIC/Mini-MUSIC - use study_id and music_timepoint
-    music_clinical_data = ClinicalData.objects.filter(
-        study_id=OuterRef("study_id"), music_timepoint=OuterRef("music_timepoint")
-    )
-
-    # Default - use study_id and sample_date
-    default_clinical_data = ClinicalData.objects.filter(
-        study_id=OuterRef("study_id"), sample_date=OuterRef("sample_date")
-    )
-
-    # Annotate with clinical data fields using conditional lookups
-    samples_with_clinical = samples_with_date.annotate(
-        crp=Case(
-            When(study_name_lower__contains="gidamps", then=Subquery(gidamps_clinical_data.values("crp")[:1])),
-            When(study_name_lower__contains="music", then=Subquery(music_clinical_data.values("crp")[:1])),
-            default=Subquery(default_clinical_data.values("crp")[:1]),
-        ),
-        calprotectin=Case(
-            When(
-                study_name_lower__contains="gidamps", then=Subquery(gidamps_clinical_data.values("calprotectin")[:1])
-            ),
-            When(study_name_lower__contains="music", then=Subquery(music_clinical_data.values("calprotectin")[:1])),
-            default=Subquery(default_clinical_data.values("calprotectin")[:1]),
-        ),
-        # Add more fields as needed using the same pattern
-        endoscopic_mucosal_healing_at_3_6_months=Case(
-            When(
-                study_name_lower__contains="music",
-                then=Subquery(music_clinical_data.values("endoscopic_mucosal_healing_at_3_6_months")[:1]),
-            )
-        ),
-        endoscopic_mucosal_healing_at_12_months=Case(
-            When(
-                study_name_lower__contains="music",
-                then=Subquery(music_clinical_data.values("endoscopic_mucosal_healing_at_12_months")[:1]),
-            )
-        ),
-    )
-
-    return samples_with_clinical
