+++ core/utils/__init__.py
@@ -1,0 +1,3 @@
+# /Users/chershiongchuah/Developer/musicsamples/core/utils/__init__.py
+# This file initializes the core.utils package.
+# It provides a namespace for utility-related modules in the core app.
